requirements.txt
fastapi[all]
uvicorn[standard]
debugpy
psycopg2
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
alembic
pytest